requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary
python-dotenv==1.0.0